% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Act.A3C}
\alias{Act.A3C}
\title{Determines which action the algorithm takes}
\usage{
Act.A3C(state, model, model.par, algo.par, game.object, worker.number, epsilon)
}
\arguments{
\item{state}{Encoded game state as given by \code{state.2.array}}

\item{model}{A model as given by \code{Setup.Neural.Network.A3C} or \code{Setup.Neural.Network.A3C.LSTM}}

\item{model.par}{A list as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}}

\item{game.object}{A Game Object (list) as defined by \code{Get.Game.Object.<NAME>}.}

\item{worker.number}{A number to identify the worker.}

\item{epsilon}{Current Value of epsilon.}
}
\description{
Internal function which is called from \code{Worker.A3C}.
}
