% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Act.Qlearning}
\alias{Act.Qlearning}
\title{Determines which action to take}
\usage{
Act.Qlearning(state, model, model.par, algo.var, game.object,
  eval.only = FALSE)
}
\description{
Determines which action to take
}
