% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPathing.R
\name{Act.QPathing}
\alias{Act.QPathing}
\title{Determines which action to take}
\usage{
Act.QPathing(state, model, model.par, algo.var, game.object,
  eval.only = FALSE, net.pointer)
}
\description{
Determines which action to take
}
