% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Act.QPredictions}
\alias{Act.QPredictions}
\title{Determines which action to take}
\usage{
Act.QPredictions(state, model, model.par, algo.var, game.object,
  eval.only = FALSE, net.pointer)
}
\description{
Determines which action to take
}
