% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Action.2.Choice.PD}
\alias{Action.2.Choice.PD}
\title{Action to Array for Prisoners Dilemma}
\usage{
Action.2.Choice.PD(output, game.object)
}
\arguments{
\item{output}{Chosen action.}

\item{game.object}{as specified by Get.Game.Object}
}
\description{
Transforms the output of a game strategy to the number of choice of the machine learning algorithm
}
