% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Action.Encoding.Info.PD}
\alias{Action.Encoding.Info.PD}
\title{Get Info of Action Encoding}
\usage{
Action.Encoding.Info.PD(game.object)
}
\description{
Returns Info of Action Encodings
}
