% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGame.R
\name{Action.Encoding.Info.Simple.Game}
\alias{Action.Encoding.Info.Simple.Game}
\title{Get Info of Action Encoding}
\usage{
Action.Encoding.Info.Simple.Game(game.object)
}
\description{
Internal Function delivering info of action encoding, as e.g. number of nodes.
}
