% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Advantage.function}
\alias{Advantage.function}
\title{Calculates N-Step Returns or weighted Temporal Difference Errors}
\usage{
Advantage.function(state, model, model.par, algo.par, game.object, queue,
  done)
}
\arguments{
\item{state}{Actual game state as returned by \code{State.Transition.<NAME>}.}

\item{model}{A model as given by \code{Setup.Neural.Network.A3C} or \code{Setup.Neural.Network.A3C.LSTM}.}

\item{model.par}{A list as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}

\item{algo.par}{A list as given by \code{Get.Def.Par.A3C}.}

\item{game.object}{A Game Object (list) as defined by \code{Get.Game.Object.<NAME>}.}

\item{queue}{Game history of the last n rounds given by \code{Convert.2.train}.}

\item{done}{Boolean specifiying wether actual episode is ended. Determines if a Bootstrap Value has to be calculated.}
}
\description{
Internal function called from \code{Worker.A3C}
}
