% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Calc.Endstate.Value.QPredictions}
\alias{Calc.Endstate.Value.QPredictions}
\title{Calculates Endstate value}
\usage{
Calc.Endstate.Value.QPredictions(net, end.pointer.pos.net, end.pointer.pos.path,
  mem.path, option = NULL)
}
\description{
Currently no option ist supported. Endstate value is calculated as average from end.pointer.pos to end of mem.path.
}
