% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Calc.Reward.QPredictions.expectedQ}
\alias{Calc.Reward.QPredictions.expectedQ}
\title{Calculate Expected Value based on action}
\usage{
Calc.Reward.QPredictions.expectedQ(net, pointer, action, option = "ignore",
  mode)
}
\description{
Needed to calculate the next Q value. If converged, should result in the same value as the respective Q.value itself.
Also outputs edge value.
}
