% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Calc.Reward.QPredictions.expectedReward}
\alias{Calc.Reward.QPredictions.expectedReward}
\title{Calculate Expected immediate Reward based on action}
\usage{
Calc.Reward.QPredictions.expectedReward(net, pointer, action,
  mode = "history")
}
\description{
Returns expected Reward based on historic data. Only used inuternally to update net as input for q values.
}
