% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Calc.Reward.QPredictions}
\alias{Calc.Reward.QPredictions}
\title{Calc.Reward.QPredictions}
\usage{
Calc.Reward.QPredictions(net, pointer, action, mode, mode.par,
  end.state = NULL, end.reward = NULL)
}
\description{
Calc.Reward.QPredictions
}
