% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Choice.2.Action.PD}
\alias{Choice.2.Action.PD}
\title{Array to Action for Prisoners Dilemma}
\usage{
Choice.2.Action.PD(output.choice, game.object)
}
\arguments{
\item{output.choice}{Number of chosen action.}

\item{game.object}{as specified by Get.Game.Object}
}
\description{
Transforms the output of an machine learning algorithm to a readable game state
}
