% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGame.R
\name{Choice.2.Action.Simple.Game}
\alias{Choice.2.Action.Simple.Game}
\title{Array to Action for Simple Game}
\usage{
Choice.2.Action.Simple.Game(output.choice, game.object)
}
\arguments{
\item{output.choice}{Number of chosen action.}

\item{game.object}{as specified by Get.Game.Object}
}
\description{
Transforms the output of an machine learning algorithm to a readable game state \cr \cr
Internal Function - There should be no need to call this function by an algorithm.
}
