% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{compare.exploration}
\alias{compare.exploration}
\title{Strategy highlighting differences in Exploration
states are "no.info", "is.nice" [after first 10 rounds], "is.mean" [if defect in first 10 rounds]}
\usage{
compare.exploration(obs, i, t, game, state = "no.info", ...)
}
\description{
Strategy highlighting differences in Exploration
states are "no.info", "is.nice" [after first 10 rounds], "is.mean" [if defect in first 10 rounds]
}
