% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Convert.2.train}
\alias{Convert.2.train}
\title{Converts stored Memory into arrays.}
\usage{
Convert.2.train(Memory)
}
\arguments{
\item{Memory}{Stored History of Worker from last n rounds.}
}
\description{
Internal function which is called from \code{Worker.A3C}.
}
