% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{counter.strat.c}
\alias{counter.strat.c}
\title{A bast answer to a testpool strategy}
\usage{
counter.strat.c(
  obs,
  i,
  t,
  game,
  P_i = 0.5,
  P_a = 0.5,
  threshold = 0.25,
  Fratio = 0.4,
  ...
)
}
\description{
Answerstrat to 'strat.c'
}
