% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Define_Graph_Gradient_Update}
\alias{Define_Graph_Gradient_Update}
\title{Graph to update Network weights}
\usage{
Define_Graph_Gradient_Update(model, model.par)
}
\arguments{
\item{model}{A Neural Network e.g. as given by \code{Setup.Neural.Network.A3c} or \code{Setup.Neural.Network.A3c.LSTM}.}

\item{model.par}{A list with parameters to set up the Network e.g. as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}
}
\description{
Defines a tensorflow graph to update Network parameters based on the gradients w.r.t to the loss function.
}
\details{
Returns a list of tensors and placeholders \itemize{
\item \strong{gradients.update} Calculate updated weights based on provided gradients.
\item \strong{vars.grads.pl} Placeholders for the gradients.
}
}
