% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Discounted.Reward.PD}
\alias{Discounted.Reward.PD}
\title{Update Score based on expected Value of reward}
\usage{
Discounted.Reward.PD(game.object, game.state, old.score, reward)
}
\description{
Returns updated score
}
