% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Encode.Game.States.PD}
\alias{Encode.Game.States.PD}
\title{Transforms List of Gamestates to std encoding form}
\usage{
Encode.Game.States.PD(game.object, game.states, expand = TRUE)
}
\description{
Returns list with states, next.states, action, reward, done
}
