% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Encoding.Harper.PD}
\alias{Encoding.Harper.PD}
\title{Encoding based on Reinforcement Learning Produces Dominant Strategies for the Iterated Prisoner's Dilemma from Harper, Knight, Jones, et. al.}
\usage{
Encoding.Harper.PD(game.state, game.object)
}
\description{
Encoding based on Reinforcement Learning Produces Dominant Strategies for the Iterated Prisoner's Dilemma from Harper, Knight, Jones, et. al.
}
