% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Encoding.last.X.rounds.PD}
\alias{Encoding.last.X.rounds.PD}
\title{Flexible Encoding Function which expects the following encoding.params:
real (TRUE or FALSE) [Should a history of own taken actions be memorized?]
rounds: How many rounds should be looked into the past?}
\usage{
Encoding.last.X.rounds.PD(game.state, game.object)
}
\description{
Flexible Encoding Function which expects the following encoding.params:
real (TRUE or FALSE) [Should a history of own taken actions be memorized?]
rounds: How many rounds should be looked into the past?
}
