% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Encoding.Manager.PD}
\alias{Encoding.Manager.PD}
\title{Internal Function to make working with different encodings easier}
\usage{
Encoding.Manager.PD(encoding.state, info = NULL)
}
\description{
Internal Function to make working with different encodings easier
}
