% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThesisFunctions.R
\name{get.against.itself.benchmark}
\alias{get.against.itself.benchmark}
\title{Payoff of strategy against itself}
\usage{
get.against.itself.benchmark()
}
\description{
Payoff of strategy against itself
}
