% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.NN.Legacy.Thesis.Basic}
\alias{Get.Def.Par.NN.Legacy.Thesis.Basic}
\title{Default Parameters Basic Neural Network as used for the thesis of Martin Kies}
\usage{
Get.Def.Par.NN.Legacy.Thesis.Basic()
}
\description{
Default Parameters Basic Neural Network as used for the thesis of Martin Kies
}
