% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.NN.Legacy.v.0.1.6}
\alias{Get.Def.Par.NN.Legacy.v.0.1.6}
\title{Default Parameters Neural Network v.0.1.6}
\usage{
Get.Def.Par.NN.Legacy.v.0.1.6()
}
\description{
Default Parameters Neural Network v.0.1.6
}
