% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Get.Def.Par.QLearning}
\alias{Get.Def.Par.QLearning}
\title{Delivers some default Parameters of Q-learning}
\usage{
Get.Def.Par.QLearning()
}
\description{
Delivers some default Parameters of Q-learning
}
