% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.QLearningPersExpPath.Legacy.ThesisOpt.XGB}
\alias{Get.Def.Par.QLearningPersExpPath.Legacy.ThesisOpt.XGB}
\title{Default Parameters QLearningPersExpPath of the thesis of Martin Kies}
\usage{
Get.Def.Par.QLearningPersExpPath.Legacy.ThesisOpt.XGB()
}
\description{
Default Parameters QLearningPersExpPath of the thesis of Martin Kies
}
