% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.QLearningPersExpPath.QLearning.Basic}
\alias{Get.Def.Par.QLearningPersExpPath.QLearning.Basic}
\title{Default Parameters normal Q-Learnung used in the thesis of Martin Kies}
\usage{
Get.Def.Par.QLearningPersExpPath.QLearning.Basic()
}
\description{
Default Parameters normal Q-Learnung used in the thesis of Martin Kies
}
