% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Get.Def.Par.QLearningPersExpPath}
\alias{Get.Def.Par.QLearningPersExpPath}
\title{Default Parameters for (improved) Q-Learning}
\usage{
Get.Def.Par.QLearningPersExpPath(setting = "ThesisOpt.XGB")
}
\arguments{
\item{setting}{String which defines the setting of the default parameters. \itemize{
  \item "ThesisOpt.XGB" - Default setting. The same parameters are used as for the final results of the thesis of Martin Kies. Uses the Gradient Boosting Variant.
  \item "ThesisOpt.RNN" - The same parameters are used as for the final results of the thesis of Martin Kies. Uses the Recurrent Neural Network (LSTM) Variant.
  \item "Legacy.v.0.1.6" - Uses the default settings of version 0.1.6 of this package
}}
}
\description{
Returns the 'algo.par' necessary for \code{\link{Setup.QLearningPersExpPath}}, \code{\link{Initialise.QLearningPersExpPath}} and \code{\link{Train.QLearningPersExpPath}}
}
