% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Get.Def.Par.QPredictions}
\alias{Get.Def.Par.QPredictions}
\title{Delivers some default Parameters of Q-Predictions}
\usage{
Get.Def.Par.QPredictions()
}
\description{
Delivers some default Parameters of Q-Predictions
}
