% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.RNN.Legacy.ThesisOpt}
\alias{Get.Def.Par.RNN.Legacy.ThesisOpt}
\title{Default Parameters Recurrent Neural Network Thesis Martin kies}
\usage{
Get.Def.Par.RNN.Legacy.ThesisOpt()
}
\description{
Default Parameters Recurrent Neural Network Thesis Martin kies
}
