% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNN.R
\name{Get.Def.Par.RNN}
\alias{Get.Def.Par.RNN}
\title{Define default Parameters of the RNN Function}
\usage{
Get.Def.Par.RNN(setting = "ThesisOpt")
}
\arguments{
\item{setting}{Specefies the to be used setting. Currently the following settings are available: \itemize{
  \item "ThesisOpt" - Default setting. The same parameters are used as for the final results of the thesis of Martin Kies.
  \item "Legacy.v.0.1.6" - Uses the default settings of version 0.1.6 of this package
}}
}
\description{
Returns a List which may be used as \code{model.par} of e.g. the function \code{Setup.QLearning()} with the following parameters:\itemize{
\item \strong{name} - Identifier of Model. Per Default \"RNN\".
\item \strong{setup} - Function which should be used to setup the RNN. Per Default \code{Setup.RNN}
\item \strong{predict} - Function which should be used to predict the RNN. Per Default \code{Predict.RNN}
\item \strong{train} - Function which should be used to train/calibrate the RNN. Per Default \code{Train.RNN}
\item \strong{hidden.nodes} - A Vector consisting of the number of Neurons in each hidden layer - e.g. c(25,10) to have two hidden layers with the first layer having 25 Neurons.
\item \strong{layer.type} - A vector consisting of the names of the type of the hidden layer. Supported are "lstm", "gru", "dense". If lstm or gru are used in a deep layer the sequence is returned.
\item \strong{activation.hidden} - A Vector defining the activation functions of the hidden layers, e.g. c(\"relu\",\"relu\"). Has to have the same number of items as \code{hidden.nodes}. Supported are e.g. relu, tanh, sigmoid and linear
\item \strong{activation.output}. Activiation function of the output layer. Supported are e.g. relu, tanh, sigmoid and linear.
\item \strong{loss}. Specifies the loss function, e.g. \'mse\'
\item \strong{optimizer}. Specifies the used optimizer. By Default Adam Optimization is used with a Learning rate of 0.001.
\item \strong{mask.value}. Which value should be used for masking?
\item \strong{epochs}. How many epochs should the RNN be trained?
\item \strong{batch.size}. Batch Size of RNN.
\item \strong{verbose}. Should the RNN give an output? 0 for no output, 1 for output for each epoch, 2 for aggregate output every other epoch.
}
}
