% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Def.Par.XGBoost.Legacy.ThesisOpt}
\alias{Get.Def.Par.XGBoost.Legacy.ThesisOpt}
\title{Default Parameters Gradient Boosting Thesis Martin Kies}
\usage{
Get.Def.Par.XGBoost.Legacy.ThesisOpt()
}
\description{
Default Parameters Gradient Boosting Thesis Martin Kies
}
