% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Game.Param.PD.Legacy.BattleOfStrategies2013.Baseline}
\alias{Get.Game.Param.PD.Legacy.BattleOfStrategies2013.Baseline}
\title{Default Parameters Game - Rules of Seminar 2013}
\usage{
Get.Game.Param.PD.Legacy.BattleOfStrategies2013.Baseline()
}
\description{
Default Parameters Game - Rules of Seminar 2013
}
