% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultFunctions.R
\name{Get.Game.Param.PD.Legacy.BattleOfStrategies2019}
\alias{Get.Game.Param.PD.Legacy.BattleOfStrategies2019}
\title{Default Parameters Game - Rules of Seminar 2019}
\usage{
Get.Game.Param.PD.Legacy.BattleOfStrategies2019()
}
\description{
Default Parameters Game - Rules of Seminar 2019
}
