% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Get.Game.Param.PD}
\alias{Get.Game.Param.PD}
\title{Standard Parameters of Repeated Prisoners Dilemma
Returns a list with parameters.}
\usage{
Get.Game.Param.PD()
}
\description{
Standard Parameters of Repeated Prisoners Dilemma
Returns a list with parameters.
}
