% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Get.Par.PD}
\alias{Get.Par.PD}
\title{Defines model parameters for 'Simple Game'}
\usage{
Get.Par.PD(game.object)
}
\arguments{
\item{game.object}{A game object as defined by Get.Game.Object.Simple.Game}
}
\description{
Public Function which might be called by algorithm functions.
Output is a list of the following structure:
\itemize{
\item input.nodes - Length of array as presented by state.2.array
\item output.nodes - Number of possible actions
}
}
