% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Initialise.A3C}
\alias{Initialise.A3C}
\title{Set changeable A3C Parameters.}
\usage{
Initialise.A3C(algo.par)
}
\arguments{
\item{algo.par}{A list as given by \code{Get.Def.Par.A3C}.}
}
\description{
Returns a structured list of lists to store the results of the learning process. \itemize{
\item \strong{algo.var.Master} Contains two lists for every worker. Memory stores the history of the worker and Graph.Output stores information of the gradient calulations. Additionally a global score is stored.
\item \strong{algo.var.worker} Contains two lists with the memory and the graph output for the given worker.
}
}
