% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Initialise.Qlearning}
\alias{Initialise.Qlearning}
\title{Set changeable model variables e.g. epsilon}
\usage{
Initialise.Qlearning(model.par)
}
\arguments{
\item{model.par}{Parameters of QLearning specification.}
}
\description{
Set changeable model variables e.g. epsilon
}
