% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Initialise.QLearningPersExpPath}
\alias{Initialise.QLearningPersExpPath}
\title{Set changeable model variables}
\usage{
Initialise.QLearningPersExpPath(
  game.object = NULL,
  algo.par,
  memory.init = NULL,
  memory.param = NULL,
  model.par = NULL
)
}
\arguments{
\item{game.object}{A Game Object as defined by \code{Get.Game.Object.<Name>}. Necessary in the case of memory intitialisation.}

\item{memory.init}{Which type of initialization should take place? It \code{NULL}, the option \code{none} is used. The following types are supported \itemize{
\item \strong{none} No initialization takes place. Memory is an empty list.
\item \strong{self.play} The other strategies play against themselves - to understand possible secret handshakes. The following \code{memory.param} are expected: \itemize{
\item \strong{no} How often should the other strategies play against themselves?
}
}
If combinations of different memories are needed, one can use the function \code{Extend.Memory.QLearningPersExpPath()}}

\item{memory.param}{Parameters necessary for the chosen \code{memory.init}.}

\item{model.par}{Parameters of model i.e. Neural Network. Currently only used for RNN.}
}
\description{
Returns a list with the following items \itemize{
\item \strong{epsilon} [only in case of epsilon.greedy] Specifies how often the Algorithm tries a random move. Initialized with \code{epsilon.start} of \code{model.par}.
\item \strong{memory} Memory of the Algorithm given past games. May be intitialised to give a starting point to learning.
}
}
