% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Initialise.QPredictions}
\alias{Initialise.QPredictions}
\title{Set changeable model variables}
\usage{
Initialise.QPredictions(game.object = NULL, algo.par, memory.init = NULL,
  memory.param = NULL)
}
\arguments{
\item{game.object}{A Game Object as defined by \code{Get.Game.Object.<Name>}. Necessary in the case of memory intitialisation.}

\item{memory.init}{Which type of initialization should take place? It \code{NULL}, the option \code{none} is used. The following types are supported \itemize{
\item \strong{none} No initialization takes place. Memory is an empty list.
\item \strong{self.play} The other strategies play against themselves - to understand possible secret handshakes. The following \code{memory.param} are expected: \itemize{
\item \strong{no} How often should the other strategies play against themselves?
}
}
If combinations of different memories are needed, one can use the function \code{Extend.Memory.QPredictions()}}

\item{memory.param}{Parameters necessary for the chosen \code{memory.init}.}

\item{model.par}{Parameters of QPredictions specification. Have to be specified and should be identical to the model.par as given to \code{Setup.QPredictions()}.}
}
\description{
Returns a list with the following items \itemize{
\item \strong{epsilon} Specifies how often the Algorithm tries a random move. Initialized with \code{epsilon.start} of \code{model.par}.
\item \strong{memory.net} A net based on the next state. Each Node holds the following information:\itemize{
\item \strong{encoding} Game state encoded for prediction algorithm.
\item \strong{depth} Current depth of tree. In the case of the Prisoners Dilemma equal to the round.
\item \strong{visited} How often have we visited this node?
\item \strong{Q.values} Q.values of this node.
\item \strong{precursors} All possibilites to get to this node
\item \strong{successors} All nodes to go to. Is a list itself with one entry for each action. Here a list with the following items exist: Number of Reached node. Number of visits.
}
}
}
