% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mystandata.R
\name{make_mystandata}
\alias{make_mystandata}
\title{make_mystandata}
\usage{
make_mystandata(
  data,
  subject_column,
  block_column,
  var_toinclude,
  var_tobenamed,
  additional_arguments
)
}
\arguments{
\item{data}{is your data frame}

\item{subject_column}{\itemize{
\item is the specific vector of subjects from you data frame (e.g., data$subject)
}}

\item{block_column}{\itemize{
\item is the specific vector of subjects from you data frame (e.g., data$block)
}}

\item{var_toinclude}{\itemize{
\item is a vector of strings with the names of data columns that should be included in the stan data
}}

\item{var_tobenamed}{\itemize{
\item is a vector of strings for how to variables should be names according to stan model (in case the data and stan has different names for the same var)
}}

\item{Ntrials_per_subject}{\itemize{
\item is a vector 1 x Nsubj with each cell including the number of available trials per subject
}}
}
\description{
aim: create data that can passed in to stan. Variables will be passed as subject x trial matrices. These matrices are padded with Inf value to have exactly the same number of trials despite missing values
}
\examples{
\dontrun{
stan_data=
make_mystandata(
data=df,
subject_column=df$subject,
block_column = df$block,
Ntrials_per_subject = 200,
var_toinclude = c("stay","ch_key","ch_card")
var_tobenamed = c("stay_card","chosen_key","chosen_card")
)
}
}
\seealso{

}
