% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Memory.Random.Play.PD}
\alias{Memory.Random.Play.PD}
\title{Generate Memory where strategies play against a random strategy}
\usage{
Memory.Random.Play.PD(game.object, algo.par)
}
\arguments{
\item{game.object}{as specified by Get.Game.Object}

\item{algo.par}{as e.g. given by Get.Def.Par.QLearning}
}
\description{
Each strategy within the game.object plays against a random strategy of the given defection probability
}
\details{
Outputs List of lists with the following elements:
\itemize{
\item state - Already encoded game state, if algo.par$mem.type=="game.encoded"
\item action - Which of the actions has been taken?
\item next.state - resulting next, encoded, state
\item reward - What did we get from transitioning to the next state?
\item done - Boolean; is the game over?
}
Public Function which might be called by algorithms.
}
