% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Model.strat.maximum.full.Ten}
\alias{Model.strat.maximum.full.Ten}
\title{A strategy to be used after model has been trained}
\usage{
Model.strat.maximum.full.Ten(obs, i, t, history.see = NULL, ...)
}
\description{
A strategy to be used after model has been trained
}
