% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posteriorplot.R
\name{my_posteriorplot}
\alias{my_posteriorplot}
\title{my_posteriorplot}
\usage{
my_posteriorplot(x, myxlabel, mycolor, myxlim, my_vline)
}
\arguments{
\item{x}{a vector you want to create a posterior from.}

\item{myxlabel}{x label}

\item{mycolor}{color of density function}

\item{myxlim}{a vector of two numbers describing the limits of the x axis.}

\item{myvline}{the value of the median or the mean.}
}
\value{
Returns a posterior plot
}
\description{
This function plots a posterior
}
\details{
This function plots a posterior from your bayesian outputs.
}
\examples{
\dontrun{
posterior_plot=(
x=param,
myxlabel="Beta1",
mycolor = "gray",
myxlim = c(0,2),
myvline= median(param)
)
}
}
\seealso{

}
