% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_raincloud.R
\name{my_raincloud}
\alias{my_raincloud}
\title{raincloud_plot}
\usage{
my_raincloud(df, x_str, y_str, xlab, ylab)
}
\arguments{
\item{x_str}{x variable for the plot as a string}

\item{y_str}{y variable for the plot as a string}

\item{xlab}{label of x}

\item{ylab}{label of y}
}
\value{
Returns a raincloud plot
}
\description{
This function plots a nice raincloud plot
}
\details{
This function plots a raincloud plot
}
\examples{
\dontrun{
raincloud_plot=
my_raincloud(
df=df,
x="working_memory",
y="stay",
xlab="Working_memory",
ylab="Probability to stay with same card")
}
}
\seealso{

}
