% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xyplot.R
\name{my_xyplot}
\alias{my_xyplot}
\title{my_xyplot}
\usage{
my_xyplot(x, y, myxlab, myylab, mycolor)
}
\arguments{
\item{x}{x variable for the plot}

\item{y}{y variable for the plot}

\item{myxlab}{label of x}

\item{myylab}{label of y}

\item{mycolor}{color of points}
}
\value{
Returns a standard plot with geom_point
}
\description{
This function plots a nice ggplot geom_point graph
}
\details{
This function standardizes the plotting procedure
}
\examples{
\dontrun{
plot=
my_xyplot(
x=working_memory,
y=stay,
myxlab="Working_memory",
myylab="Probability to stay with same card",
mycolor = "green")
}
}
\seealso{

}
