% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{NN.strat.full.zero}
\alias{NN.strat.full.zero}
\title{A strategy to be used after model has been trained}
\usage{
NN.strat.full.zero(obs, i, t, history.see = NULL, ...)
}
\description{
A strategy to be used after model has been trained
}
