% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{NN.strat}
\alias{NN.strat}
\title{The actual strategy after model has been trained}
\usage{
NN.strat(obs, i, t, history.see = NULL, ...)
}
\description{
Does not work for itself - the "model" variable has to be specified beforehand.
}
