% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{NN.strat.Slim.TenTen}
\alias{NN.strat.Slim.TenTen}
\title{A strategy to be used after model has been trained}
\usage{
NN.strat.Slim.TenTen(obs, i, t, history.see = NULL, net.pointer = 1, ...)
}
\description{
A strategy to be used after model has been trained
}
