% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{phases}
\alias{phases}
\alias{phases}
\title{A student strategy}
\usage{
phases(obs, i, t, equal = TRUE, offender = FALSE, victim = FALSE, ...)

phases(obs, i, t, equal = TRUE, offender = FALSE, victim = FALSE, ...)
}
\description{
A student strategy

A student strategy
}
