% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpfullFunctions.R
\name{PID.controller}
\alias{PID.controller}
\title{PID controller, which generates ouptput based on the error}
\usage{
PID.controller(
  err.vec,
  Kp,
  Ki,
  Kd,
  Kp.db = 1,
  Kp.disc = 0.75,
  Ki.db = Inf,
  Kd.db = 1,
  Kd.disc = 0.75,
  Ki.disc = 0.75
)
}
\arguments{
\item{err.vec}{A vector with the error terms (discrepancies of the result and the value)}

\item{Kp}{Factor measuring how strong we want to turn based on proportial gain}

\item{Ki}{Factor measuring how strong we want to turn based on integral gain}

\item{Kd}{Factor measuring how strong we want to turn based on derivative gain}

\item{Kp.db}{How many data points should be used to calculate the proportional effect? By default the standard procedure is used, e.g. only the last data point. If there is a lot of noise in the environment it might be prudent to use a higher number.}

\item{Kp.disc}{When calculating the proportional effect based on multiple data.points with Kp.db it might be sensible to weight more recent data points higher. This parameter determines the discount factor.}

\item{Ki.db}{How many data points should be used to calculate the integral? By default the standard procedure is used, e.g. all data points. If there are strong changes in the environment it might be prudent to use a lower number.}

\item{Kd.db}{How many data points should be used to calculate the derivative? By default the standard procedure is used, e.g. the change between the last two data points. If there is a lot of noise one might want to use a higher number of points.}

\item{Kd.disc}{When calculating the derivative based on multiple data.points with Kd.db it might be sensible to weight more recent changes higher. This parameter determines the discount factor to determine the weights of former changes.}

\item{Ki.disc}{When calculating the integral based on multiple data.points with Kp.db it might be sensible to weight more recent data points higher. This parameter determines the discount factor.}
}
\description{
PID controller, which generates ouptput based on the error
}
