% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Play.On.Strategy.QLearningPersExpPath}
\alias{Play.On.Strategy.QLearningPersExpPath}
\title{Play the game based on strategy}
\usage{
Play.On.Strategy.QLearningPersExpPath(model, model.surp = NULL,
  model.fam = NULL, model.par, model.par.surp = NULL,
  model.par.fam = NULL, algo.par, algo.var, game.object,
  best.guess.episode = FALSE, eval.only = TRUE, block.type = NULL)
}
\description{
Play the game based on strategy
}
