% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Predict.Neural.Network.A3C}
\alias{Predict.Neural.Network.A3C}
\title{Predict Neural Network}
\usage{
Predict.Neural.Network.A3C(model, model.par, state, Indicator = NULL)
}
\arguments{
\item{model}{A Neural Network e.g. as given by \code{Setup.Neural.Network.A3c} or \code{Setup.Neural.Network.A3c.LSTM}.}

\item{model.par}{A list with parameters to set up the Network e.g. as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}

\item{state}{An encoded game state.}

\item{Indicator}{Indicates which value should be returned. Could be either: \itemize{
\item \strong{"Value"} Only the predicted value of the state is returned.
\item \strong{"Prob"} Only the action probabilites of the given state are returned.
\item \strong{NULL} Both are retured.
}}
}
\description{
Predict action probabilities and state values based on a neural network.
}
