% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{prep.data.4.shiny}
\alias{prep.data.4.shiny}
\title{Prepare Worker Memory to visualize with shiny}
\usage{
prep.data.4.shiny(threads.data, choices, actions)
}
\arguments{
\item{threads.data}{Memory of experience of one worker.}

\item{choices}{Possible resulting choices as a string vector. In the IPD game this would be c("CC", "CD","DC","DD").}

\item{actions}{Actions space of the specific game [Currently only IPD is supported.] as a string vector.}
}
\description{
Memory of one worker is transformed so it can be as input to the shiny app.
}
\details{
Returns a list with following items \itemize{

\item \strong{memory.mat} The experience of a worker stored in a numeric matrix.
\item \strong{graph.mat} Network loss data stored in a numeric matrix.
\item \strong{help.data} Additional data used in the shiny app.
}
}
