% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalk.R
\name{randomwalk}
\alias{randomwalk}
\title{randomwalk_example}
\usage{
randomwalk(
  Narms = 2,
  Ntrials = 100,
  tau = 0.02,
  rho = 0,
  upper.bound = 0.85,
  lower.bound = 0.15,
  plot = TRUE,
  save_csv = TRUE
)
}
\arguments{
\item{Narms}{Number of bandit arms, Default: 2}

\item{Ntrials}{Number of trials, Default: 100}

\item{tau}{Standard deviation for the noise normal distribution of each arm, Default: 0.02}

\item{rho}{True correlation between all arms. default should be zero, Default: 0}

\item{upper.bound}{Maximal limit of reward probability, Default: 0.85}

\item{lower.bound}{Minimal limit of reward probability, Default: 0.15}

\item{plot}{Whether to plot the bandits}

\item{save_csv}{Whether to save the csv for your experiment}
}
\value{
Returns a matrix (Ntrials X Narms) showing the reward probabilities in each trial.
}
\description{
This function generates random reward probabilities for the whole experiment
}
\details{
The randomwalk function is responsible for generating the rewards in your experiment.
Each arm has a different reward probability and it drifts along the experiment. The rational behind the drift is to maintain learning.
}
\examples{
\dontrun{
{
R=
randomwalk(Narms=4,
          Ntrials=50,
          tau            =.02,#standard deviation for the noise normal distribution of each arm
          rho            =0, #true correlation between all arms. default should be zero
          upper.bound   =0.85,
          lower.bound   =0.15)
#check correlations
cor(R)
 }
}
}
\seealso{
\code{\link[MASS]{mvrnorm}}
}
