% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{redim.state}
\alias{redim.state}
\title{Change dimensionality of the state array.}
\usage{
redim.state(state)
}
\arguments{
\item{state}{An ecoded array of game states}
}
\description{
Helper funciton to add a batch dimensionalty needed for the LSTM-Neural Network.
}
