% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Replay.QLearningPersExpPath}
\alias{Replay.QLearningPersExpPath}
\title{Train model of Q learning}
\usage{
Replay.QLearningPersExpPath(evaluator, model.par, model.par.surp = NULL,
  model.par.fam = NULL, algo.par, algo.var, game.object)
}
\arguments{
\item{model.par.surp}{If \code{NULL}, no suprise model is generated/updated.}

\item{model.par.fam}{If \code{NULL}, no familiarity model is generated/updated.}
}
\description{
Train model of Q learning
}
