% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPathing.R
\name{Replay.QPathing}
\alias{Replay.QPathing}
\title{Train model of Q Pathing}
\usage{
Replay.QPathing(model, model.par, algo.par, algo.var, game.object)
}
\description{
As QPathing updates all new paths directly Replay only trains the prediction model.
}
