% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Replay.QPredictions}
\alias{Replay.QPredictions}
\title{Train model of Q Pathing}
\usage{
Replay.QPredictions(model, model.par, algo.par, algo.var, game.object)
}
\description{
As QPredictions updates all new paths directly Replay only trains the prediction model.
}
