% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_values.R
\name{sample_values}
\alias{sample_values}
\title{simulate_values}
\usage{
sample_values(
  population_size,
  mean_group1,
  mean_group2,
  sd_group1,
  sd_group2,
  interaction_difference,
  independent_var_name_1,
  independent_var_name_2,
  dependent_var_name
)
}
\arguments{
\item{population_size}{Size of the population you simulate}

\item{mean_group1}{mean of first condition}

\item{mean_group2}{mean of second condition}

\item{sd_group1}{sd of first condition}

\item{sd_group2}{sd of second condition}

\item{interaction_difference}{the interaction in y values}

\item{independent_var_name_1}{name of variable 1}

\item{independent_var_name_2}{name of variable 2}

\item{dependent_var_name}{name of dependent variable}
}
\value{
Returns a df showing the values of the population in all the conditions
}
\description{
This function simulates values of 4 groups by 2 conditions
}
\details{
This function is the first step in the power simulation pipeline
}
\examples{
\dontrun{
{
df =
sample_values(population_size=10000,
mean_group1=0.54,
mean_group2=0.51,
sd_group1=0.04,
sd_group2=0.02,
interaction_difference=0.02,
independent_var_name_1 = "reward",
independent_var_name_2 = "load",
dependent_var_name = "stay_key")
 }
}
}
\seealso{

}
