% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{screams.in.space}
\alias{screams.in.space}
\title{A grad student strategy}
\usage{
screams.in.space(obs, i, t, preObs = c("C", "C"), ans = "C", preAns = "C",
  count.devious = 0, clueless = 0, count.obserr = 0, ...)
}
\description{
A grad student strategy
}
