% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Setup.Neural.Network.A3C.LSTM}
\alias{Setup.Neural.Network.A3C.LSTM}
\title{Setup a  Neural Network with an LSTM-Layer for the A3C-implementation.}
\usage{
Setup.Neural.Network.A3C.LSTM(model.par, game.object)
}
\arguments{
\item{model.par}{A list with parameters to set up the Network e.g. as given by \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}

\item{game.par}{A list with parameters of the game.}
}
\description{
Returns a Recurrent Neural Network based on keras.
}
