% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Setup.Neural.Network.A3C}
\alias{Setup.Neural.Network.A3C}
\title{Setup a Feed-Forward Neural Network for the A3C-implementation.}
\usage{
Setup.Neural.Network.A3C(model.par, game.object)
}
\arguments{
\item{model.par}{A list with parameters to set up the Network e.g. as given by \code{Get.Def.Par.Neural.Network.A3C}.}

\item{game.par}{A list with parameters of the game.}
}
\description{
Returns a Feed-Forward Neural Network based on keras.
Important: At the moment this does only support discrete action spaces. For the continous case use \code{Setup.Neural.Network.A3C.LSTM}
}
