% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetwork.R
\name{Setup.Neural.Network}
\alias{Setup.Neural.Network}
\title{Setup a Neural Network}
\usage{
Setup.Neural.Network(model.par, game.par)
}
\arguments{
\item{model.par}{Parameters of Neural Network e.g. given by \code{Get.Def.Par.Neural.Network}}

\item{game.par}{Parameters of Game. Used are \itemize{
\item input.nodes - Number of Input Nodes
\item output.nodes - Number of Actions
}}
}
\description{
Setup the Neural Network in keras to work with it. Returns a keras stile Neural Network
}
