% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Setup.QLearning}
\alias{Setup.QLearning}
\title{Sets up a model based on model parameters}
\usage{
Setup.QLearning(game.object, model.par = NULL)
}
\description{
Sets up a model based on model parameters
}
