% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Setup.QLearningPersExpPath}
\alias{Setup.QLearningPersExpPath}
\title{Sets up a model based on model parameters}
\usage{
Setup.QLearningPersExpPath(game.object, algo.par = NULL, model.par,
  model.par.surp = NULL, model.par.fam = NULL)
}
\arguments{
\item{game.object}{Game Object as defined by \code{Get.Game.Object.<NAME>}.}

\item{model.par}{Model parameters. If \code{NULL}, the function \code{Get.Def.Par.QLearningPersExpPath()} is called.}

\item{model.par.surp}{Model parameters for surprise model. If \code{NULL}, we assume that we do not want to use a surprise model.}

\item{model.par.fam}{Model parameters for the model which tries to gauge familiarity with the specific state/action combination. If \code{NULL}, we assume that no familiarity model is needed.}
}
\description{
Sets up a model based on model parameters
}
