% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPathing.R
\name{Setup.QPathing}
\alias{Setup.QPathing}
\title{Q-Pathing is rather similar to Q-learning but we have stronger assumptions:\itemize{
\item We only receive our reward at the end.
\item We update instantly.
\item We use our predictive model only in those cases, where we have not yet found information.
\item We have a single starting state.
}}
\usage{
Setup.QPathing(game.object, algo.par = NULL, model.par)
}
\arguments{
\item{game.object}{Game Object as defined by \code{Get.Game.Object.<NAME>}.}

\item{model.par}{Model parameters. If \code{NULL}, the function \code{Get.Def.Par.QPathing()} is called.}
}
\description{
Q-Pathing is rather similar to Q-learning but we have stronger assumptions:\itemize{
\item We only receive our reward at the end.
\item We update instantly.
\item We use our predictive model only in those cases, where we have not yet found information.
\item We have a single starting state.
}
}
