% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPredictions.R
\name{Setup.QPredictions}
\alias{Setup.QPredictions}
\title{Q-Predictions is rather similar to Q-learning but we use a multilayer approach:\itemize{
\item We assume that even though the algorithm doesn't have complete information at runtime it may be used for training
\item A predictive neural net is used to calculate the next move of the opponent to use monte carlo studies.
\item Assumption: We have a single starting state.
}}
\usage{
Setup.QPredictions(game.object, algo.par = NULL, model.par)
}
\arguments{
\item{game.object}{Game Object as defined by \code{Get.Game.Object.<NAME>}.}

\item{model.par}{Model parameters. If \code{NULL}, the function \code{Get.Def.Par.QPredictions()} is called.}
}
\description{
Q-Predictions is rather similar to Q-learning but we use a multilayer approach:\itemize{
\item We assume that even though the algorithm doesn't have complete information at runtime it may be used for training
\item A predictive neural net is used to calculate the next move of the opponent to use monte carlo studies.
\item Assumption: We have a single starting state.
}
}
