% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpfullFunctions.R
\name{smooth.average}
\alias{smooth.average}
\title{Calculates a sensible moving average based on smoothing parameter}
\usage{
smooth.average(input, smooth = 5, show = "value")
}
\arguments{
\item{smooth}{How much should be smoothed. This is the total value of used data points, so (smooth-1)/2 is smoothed in either direction.}

\item{show}{Either 'value' for the actual value or 'sd' for a running of the averaged Standarddeviation within the smoothing range}
}
\description{
Calculates a sensible moving average based on smoothing parameter
}
