% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{squishy.the.octopus}
\alias{squishy.the.octopus}
\title{A student strategy}
\usage{
squishy.the.octopus(obs, i, t, preObs = c("C", "C"), ans = "C",
  preAns = "C", defectTotal = 0, defectConsecutive = 0,
  stopLoss = FALSE, ...)
}
\description{
A student strategy
}
