% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{State.2.Array.PD}
\alias{State.2.Array.PD}
\title{State to Array for Prisoners Dilemma}
\usage{
State.2.Array.PD(game.state, game.object)
}
\arguments{
\item{game.state}{Game State in human readable form.}

\item{game.object}{as specified by Get.Game.Object
Public Function which might be called by algorithms.}
}
\description{
Transforms Game State to readable array. Here a lot of optimization might take place, as different algorithms might like different encodings.
}
