% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGame.R
\name{State.Transition.Simple.Game}
\alias{State.Transition.Simple.Game}
\title{Get next State of Simple Game}
\usage{
State.Transition.Simple.Game(game.state, action, game.object)
}
\arguments{
\item{game.state}{A complete game state, non encoded}

\item{action}{The choice of action.}

\item{game.object}{as specified by Get.Game.Object}
}
\description{
Outputs List with three elements:
\itemize{
\item next.state - next game state (in Complete Form)
\item reward - What did we get from transitioning to the next state?
\item game.finished - Boolean; is the game over?
}
Public Function which might be called by algorithms.
}
