% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{strat.d}
\alias{strat.d}
\title{A testpool strategy}
\usage{
strat.d(
  obs,
  i,
  t,
  game,
  Cooperation.Count = 0,
  Consecutive.Cooperations = 0,
  Consecutive.Defections = 0,
  ...
)
}
\description{
A testpool strategy
}
