% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{strat3}
\alias{strat3}
\title{A student strategy}
\usage{
strat3(obs, i, t, counter = 0, standi = "G", obsl = c("Start"),
  obslown = c("Start"), ansl = c("Start"), ...)
}
\description{
A student strategy
}
