% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDstrategies.R
\name{strat4}
\alias{strat4}
\title{A student strategy}
\usage{
strat4(obs, i, t, itsyou = 0, still.defect = 3, still.cooperate = 3,
  haveto.cooperate = 0, haveto.defect = 3, endto.cooperate = 3,
  knowwho = 0, ...)
}
\description{
A student strategy
}
