% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Train.A3c}
\alias{Train.A3c}
\title{Use the A3C algorithm to train a model}
\usage{
Train.A3c(model.par, algo.par, algo.var, game.object, episodes.max)
}
\arguments{
\item{model.par}{A list as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}

\item{algo.par}{A list as given by \code{Get.Def.Par.A3C}.}

\item{algo.var}{A list as given by \code{Initialise.A3C}.}

\item{game.object}{A Game Object (list) as defined by \code{Get.Game.Object.<NAME>}.}

\item{episodes.max}{Specifies how many episodes should be played in total.}
}
\description{
Returns a list as given by \code{Initialise.A3C} with stored experiences for each worker.
}
