% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetwork.R
\name{Train.Neural.Network}
\alias{Train.Neural.Network}
\title{Train Neural Network}
\usage{
Train.Neural.Network(model, model.par, x_train, y_train)
}
\arguments{
\item{model}{A Neural Network e.g. given by \code{Setup.Neural.Network}}

\item{model.par}{Parameters of Neural Network, e.g. given by \code{Get.Default.Neural.Network}}

\item{x_train}{A Matrix with as much columns as input parameters in the encoding.}

\item{y_train}{A Matrix with as much columns as output parameters (e.g. action parameters). For each state (x_train)/action combination(column of y_train) this value determines the target.}
}
\description{
Trains a neural network and prints some helpful statistics. Returns a trained model.
}
