% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPathing.R
\name{Train.QPathing}
\alias{Train.QPathing}
\title{Train a model based on Q-Learning}
\usage{
Train.QPathing(model, model.par, algo.par, algo.var, game.object, episodes,
  eval.only = FALSE, start.w.training = TRUE)
}
\description{
Train a model based on Q-Learning
}
