% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNN.R
\name{Train.RNN}
\alias{Train.RNN}
\title{Train RNN}
\usage{
Train.RNN(model, model.par, x_train, y_train, action)
}
\arguments{
\item{model}{A RNN e.g. given by \code{Setup.RNN}}

\item{model.par}{Parameters of RNN, e.g. given by \code{Get.Default.RNN}}

\item{x_train}{A 3D array in the form c(minibatch.size, max.length.sequence, outputs)}

\item{y_train}{In the case of !single.dimensional: A Matrix with as much columns as output parameters (e.g. action parameters). For each state (x_train)/action combination(column of y_train) this value determines the target. If single.dimensional this is a simple vector with the result and the function will build a sufficient x/y pair based on action.}

\item{action}{Necessary if not single dimensional - actions which led to y based on x.}
}
\description{
Trains a RNN and prints some helpful statistics. Returns a trained model.
}
