% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Update.Evaluator.QLearningPersExpPath}
\alias{Update.Evaluator.QLearningPersExpPath}
\title{Controlled Copying of Models}
\usage{
Update.Evaluator.QLearningPersExpPath(evaluator = evaluator,
  new.model = NULL, new.model.surp = NULL, new.model.fam = NULL,
  which.update = "both", model.par)
}
\arguments{
\item{evaluator}{Evaluation consisting of the two models}

\item{new.model}{Model with which we want to update. If \code{NULL}, no update is done.}

\item{new.model.surp}{If not \code{NULL}, the surprise model is updated accordingly.}

\item{new.model.fam}{If not \code{NULL}, the familiarity model is updated accordingly.}

\item{which}{May be `both`, `cur` or `best` and updates respectively.}
}
\description{
Returns a new, updated evaluator
}
