% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QlearningPersExpPath.R
\name{Update.Memory.QLearningPersExpPath}
\alias{Update.Memory.QLearningPersExpPath}
\title{Add historic Q-Values and Curiosity to memory}
\usage{
Update.Memory.QLearningPersExpPath(
  algo.var,
  new.mem,
  game.object,
  algo.par,
  best.effort = NULL
)
}
\description{
Add historic Q-Values and Curiosity to memory
}
