% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynchronousAdvantageActorCritic.R
\name{Worker.A3C}
\alias{Worker.A3C}
\title{Defines an Agent based on the A3C-Algorithm}
\usage{
Worker.A3C(worker.number, game.state, model.par, algo.par, algo.var.worker,
  game.object, weights, Worker.Episode, model)
}
\arguments{
\item{worker.number}{Number to identify the given worker.}

\item{game.state}{Game State from the last round.}

\item{model.par}{A list as given by \code{Get.Def.Par.Neural.Network.A3C} or \code{Get.Def.Par.Neural.Network.A3C.LSTM}.}

\item{algo.par}{A list as given by \code{Get.Def.Par.A3C}.}

\item{algo.var.worker}{A list as given by \code{Initialise.A3C}.}

\item{game.object}{A Game Object (list) as defined by \code{Get.Game.Object.<NAME>}.}

\item{weights}{Network weights from the Master Network.}

\item{Worker.Episode}{The actual Episode of the Worker.}

\item{model}{A Model as given by \code{Setup.Neural.Network.A3C} or \code{Setup.Neural.Network.A3C.LSTM}}
}
\description{
Internal function which is called from \code{Train.A3C}.
}
\details{
Returns a list with the following items \itemize{
\item \strong{training} A list containing calculated gradients.
\item \strong{reward} Last obtained reward during the current n-steps.
\item \strong{finished} Boolean specifiying wether episode is finished or not.
\item \strong{algo.var.Worker} List containing the stored experience of the given worker.
}
}
