% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Get.Game.Object.PD}
\alias{Get.Game.Object.PD}
\title{Get Game Object which fully defines Prisoners Dilemma.}
\usage{
Get.Game.Object.PD(encoding.state = "Main.real",
  encoding.action = NULL, encoding.params = NULL,
  eval.strategy = Model.strat.Main.real.Exp.Path,
  basic.name.eval.model = "model.strat.global",
  basic.name.eval.model.par = "model.par.strat.global",
  game.setting = "BattleOfStrategies2013.Baseline",
  strats = "tit.for.tat")
}
\arguments{
\item{encoding.state}{Which feature selection should be used to encode the game? Default case is the main encoding of the thesis of Martin Kies, Main.real}

\item{encoding.action}{Which method should be used to encode the action? Currently supported:
\itemize{
\item main - [C,D]
}}

\item{encoding.params}{Insofar the \code{encoding.state} is parametrizable, e.g. "Model.strat.RNN.TimeSeries.flexible" or "Encoding.last.X.rounds.PD", here the encoding specific parameters are designed.}

\item{eval.strategy}{The name of the strategy used if one wants to evaluate the strategy with the package StratTourn or if self-play is used (i.e. if set to play a tournament with the strategy itself being a viable participant). By default "Model.strat.Main.real.Exp.Path" is used, which is designed to work with the encoding "Main.real".}

\item{basic.name.eval.model}{Global name under which the name of the model is to be saved.}

\item{basic.name.eval.model.par}{Global name under which the name of the parameter list of the model is to be saved.}

\item{game.setting}{Default settings of the game. By default the setting "BattleOfStrategies2013.Baseline" is used, which is the relevant one for the thesis of Martin Kies. For more information regarding the implemented settings see \code{link{Get.Game.Param.PD}}.}

\item{strats}{The strategies against which one wants to play. If several strategy names are given, each episode a random strategy is chosen. The code name "self" implements self play.}
}
\description{
Get Game Object which fully defines Prisoners Dilemma.
}
