% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Get.Game.Object.PD}
\alias{Get.Game.Object.PD}
\title{Get Game Object which fully defines Prisoners Dilemma.}
\usage{
Get.Game.Object.PD(encoding.state = NULL, encoding.action = NULL)
}
\arguments{
\item{encoding.state}{Which method should be used to encode the game? Currently supported:
\itemize{
\item main \itemize{
\item [1] Bit - See C (other)
\item [2] Bit - See D (other)
\item[3] Bit - See C (me)
\item[4] Bit - See D (other)
\item[5] Int - Round/100
\item[6] Bit - is first round
\item[7] Int - number of D of other per round
\item[8] Int - number of D of me per round
}
}}

\item{encoding.action}{Which method should be used to encode the action? Currently supported:
\itemize{
\item main - [C,D]
}}
}
\description{
Get Game Object which fully defines Prisoners Dilemma.
}
