% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrisonersDilemmaStratTourn.R
\name{Get.Game.Param.PD}
\alias{Get.Game.Param.PD}
\title{Defines the game environment of the IPD}
\usage{
Get.Game.Param.PD(
  setting = "BattleOfStrategiesThesis.Baseline",
  strats = "tit.for.tat"
)
}
\arguments{
\item{setting}{Here, one defines the setting which should be used as a default. The following settings are currently implemented: \itemize{
 \item "BattleOfStrategiesThesis.Baseline" - The setting as defined in the thesis of Martin Kies. 15% Prob. of mistaking a cooperation as a defection. \eqn{delta} 0.95. T.max 60 for numerical purposes as default.  
 \item "BattleOfStrategies2019" - The setting as defined the seminar "Fortgeschrittene Analyse von Kooperation mit Spieltheorie und Simulation aka Battle of Strategies" of winter term 2019/2020. 25% Prob. of mistaking an observation in either direction. \eqn{delta} 0.985. T.max 60 for numerical purposes as default.  
}}

\item{strats}{Names of strategies which are to be optimized against. If several strategies are given it is random against which strategy the game is played with each episode. Playing against itself is indicated by "self". Note that this function sets the strategy to NA if "self" is wished and the respective encoding strategy has to be manually filled in.}
}
\description{
Returns a list with the parameters.
}
