% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/my_posteriorplot.R
\name{my_posteriorplot}
\alias{my_posteriorplot}
\title{my_posteriorplot}
\usage{
my_posteriorplot(x, myxlabel, mycolor, myxlim, my_vline)
}
\arguments{
\item{x}{a vector you want to create a posterior from.}

\item{myxlabel}{x label}

\item{mycolor}{color of density function}

\item{myxlim}{a vector of two numbers describing the limits of the x axis.}

\item{myvline}{the value of the median or the mean.}
=======
% Please edit documentation in R/my_posterior_plot.R
\name{my_posteriorplot}
\alias{my_posteriorplot}
\title{my_posterior_plot}
\usage{
my_posteriorplot(model, variable_name, xlabel, mycolor)
}
\arguments{
\item{model}{a model you want to create a posterior from.}

\item{variable_name}{the variable you want to model}

\item{xlabel}{the label for the variable}

\item{mycolor}{color of the density function}
>>>>>>> f6c651438e4968a9509e307fd7bcb68a118a30e0
}
\value{
Returns a posterior plot
}
\description{
This function plots a posterior
}
\details{
<<<<<<< HEAD
This function plots a posterior from your bayesian outputs.
=======
This function plots a posterior from your Bayesian outputs.
>>>>>>> f6c651438e4968a9509e307fd7bcb68a118a30e0
}
\examples{
\dontrun{
posterior_plot=(
<<<<<<< HEAD
x=param,
myxlabel="Beta1",
mycolor = "gray",
myxlim = c(0,2),
myvline= median(param)
=======
model=model,
variable_name="reward_oneback",
xlabel = "Reward Oneback"
mycolor = "gray",
>>>>>>> f6c651438e4968a9509e307fd7bcb68a118a30e0
)
}
}
\seealso{

}
