% Generated by roxygen2: do not edit by hand
\name{my_posteriorplot}
\alias{my_posteriorplot}
\title{my_posteriorplot}
\usage{
my_posteriorplot(model, variable_name, xlabel, mycolor)
}
\arguments{
\item{model}{a model you want to create a posterior from.}

\item{variable_name}{the variable you want to model}

\item{xlabel}{the label for the variable}

\item{mycolor}{color of the density function}
}
\value{
Returns a posterior plot
}
\description{
This function plots a posterior
}
\details{
This function plots a posterior from your bayesian outputs.
}
\examples{
\dontrun{
my_posteriorplot=(
model=model,
variable_name="reward_oneback",
xlabel = "Reward Oneback"
mycolor = "gray"
)
}
}
\seealso{

}
