% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLR-package.R
\docType{package}
\name{RLR-package}
\alias{RLR}
\alias{RLR-package}
\title{RLR - All you Need for Your Reinforcement Learning Experiment}
\description{
This package provides helper function necessary for RL research.
}
\seealso{
\href{https://www.shahar-lab.com/}{Computational Clinical Psychology Lab}
}
\author{
Ido Ben Artzi, \email{ido.iba@gmail.com} ; Nitzan Shahar, \email{shahar.nitzan@gmail.com}
}
