% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLR-package.R
\docType{package}
\name{RLR-package}
\alias{RLR}
\alias{RLR-package}
\title{RLR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ido Ben Artzi \email{ido.iba@gmail.com} (\href{https://orcid.org/
0000-0001-5971-8613}{ORCID})

}
\keyword{internal}
