% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Setup.QLearning}
\alias{Setup.QLearning}
\title{Sets up a model based on model parameters}
\usage{
Setup.QLearning(game.object, model.par = NULL)
}
\arguments{
\item{game.object}{Game Object as defined by \code{Get.Game.Object.<NAME>}.}

\item{model.par}{Model parameters. If \code{NULL}, the function \code{Get.Def.Par.QLearning()} is called.}
}
\description{
Sets up a model based on model parameters
}
