% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning.R
\name{Train.QLearning}
\alias{Train.QLearning}
\title{Train a model based on Q-Learning}
\usage{
Train.QLearning(model, model.var, model.par, episodes, game.object)
}
\description{
Train a model based on Q-Learning
}
